The class declaration is moved to a header.
